Update tview to use a coroutine-per-device and add recovery logic
This lets the setup be written in a linear fashion rather than as a
state machine, and makes it more robust at the same time.

If it looks like we're getting binary garbage over the diagnostic
link, attempt to resynchronize by looking for a frame where no data is
received.  On Windows, while resizing the window, either pyside or
Windows doesn't run the event loop at all, which can cause data to be
discarded.  With this fix, a single line of garbage will still show
up, but at least it will likely recover eventually.

Hopefully resolves #4